cgal client added to clean results + minor changes

+++ 00_build_city3d.sh
@@ -5,10 +5,13 @@
 BUILD_DIR="Release"
 CGAL_VERSION="5.2.2"
 
-# Download CGAL (to be used as header only)
-wget "https://github.com/CGAL/cgal/releases/download/v$CGAL_VERSION/CGAL-$CGAL_VERSION.tar.xz" .
-tar -xf CGAL-$CGAL_VERSION.tar.xz
-rm CGAL-$CGAL_VERSION.tar.xz
+if [ ! -d "CGAL-$CGAL_VERSION" ]
+then
+    # Download CGAL (to be used as header only)
+    wget "https://github.com/CGAL/cgal/releases/download/v$CGAL_VERSION/CGAL-$CGAL_VERSION.tar.xz" .
+    tar -xf CGAL-$CGAL_VERSION.tar.xz
+    rm CGAL-$CGAL_VERSION.tar.xz
+fi
 
 # Recreate build dir
 if [ -d "$BUILD_DIR" ]
@@ -16,7 +19,7 @@ then
     rm -rf $BUILD_DIR
 fi
 
-# Compile city3d shape executables
+# Compile City3D release executables
 mkdir $BUILD_DIR
 cd $BUILD_DIR
 cmake -DCMAKE_BUILD_TYPE=Release ..
